Move bibtexutils to providers/bibtexformatterlib
