+++ src/providers/bibtexformatterlib/bibtexutils.ts
